requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ratelimit==4.1.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pyotp==2.9.0
sqlparse==0.5.0
typing_extensions==4.12.0
